# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateShareIn": {
        "description": "Body for POST /documents/{slug}/share.",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO datetime in UTC. Past values are rejected (422).",
            "title": "Expires At"
          },
          "notify_emails": {
            "description": "선택적으로 share-link 안내 메일을 보낼 수신자 목록. 각 항목에 대해 best-effort SMTP 발송 — 실패해도 token 생성은 성공한다.",
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Notify Emails",
            "type": "array"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 4,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional password gate; hashed with argon2.",
            "title": "Password"
          }
        },
        "title": "CreateShareIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "MX 사업부 위키 백엔드. DocumentJSON v1.0 (`packages/shared/schemas/document.json`) 을 단일 진실 공급원(SSOT) 으로 사용한다.\n\n**핵심 흐름**: `POST /api/v1/documents` 한 번으로 본문 검증 → 섹션 번호 부여 → 위키링크/태그/용어집/감사로그/검색인덱스가 모두 자동 갱신된다.\n\n**Conditional write**: 본문 수정(PUT/PATCH) 은 `If-Match` 헤더 필수 (낙관적 동시성). GET 응답의 `ETag` 를 그대로 사용하면 된다.",
    "title": "MX White Paper API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{slug}": {
      "delete": {
        "description": "status='archived' 로 표시. 검색 인덱스에서 제거되며 list 결과에 노출되지 않는다.",
        "operationId": "delete_document_api_v1_documents__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-MXWP-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Mxwp-User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서 archive (soft delete)",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "description": "DocumentJSON 본문을 그대로 반환. 응답 헤더 `ETag` 를 PUT/PATCH 의 If-Match 에 그대로 사용.",
        "operationId": "get_document_api_v1_documents__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Document Api V1 Documents  Slug  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "단일 문서 조회 (ETag 포함)",
        "tags": [
          "documents"
        ]
      },
      "put": {
        "description": "DocumentJSON 본문을 통째로 교체한다. `If-Match` 헤더에 현재 ETag 를 보내야 한다 (낙관적 동시성).\n\n변경 시 links/tags/glossary/검색 인덱스가 모두 재동기화되며, `metadata.part` 한글 이름도 자동 해석된다.",
        "operationId": "replace_document_api_v1_documents__slug__put",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "X-MXWP-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Mxwp-User"
            }
          },
          {
            "in": "header",
            "name": "X-MXWP-Change-Log",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Mxwp-Change-Log"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replace Document Api V1 Documents  Slug  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서 전체 교체 (DocumentJSON 본문 + If-Match 필요)",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{slug}/share": {
      "get": {
        "description": "revoked_at IS NULL 인 항목만 반환. view_count, expires_at 포함.",
        "operationId": "list_share_links_api_v1_documents__slug__share_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Share Links Api V1 Documents  Slug  Share Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "활성 공유 링크 목록 (editor+)",
        "tags": [
          "sharing"
        ]
      },
      "post": {
        "description": "문서 slug 에 대한 새 share token 을 발급한다. `expires_at` 미지정 시 무기한, `password` 지정 시 argon2 해시로 저장되며 공개 GET 에서 반드시 일치해야 한다.",
        "operationId": "create_share_link_api_v1_documents__slug__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Share Link Api V1 Documents  Slug  Share Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "공개 공유 링크 생성 (editor+)",
        "tags": [
          "sharing"
        ]
      }
    },
...
    "/api/v1/share/{token}": {
      "delete": {
        "operationId": "revoke_share_link_api_v1_share__token__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "공유 링크 revoke (creator 또는 admin)",
        "tags": [
          "sharing"
        ]
      },
      "get": {
        "description": "토큰이 유효하면 `GET /documents/{slug}` 와 동일한 envelope 을 반환한다. 비밀번호가 설정된 링크는 `password` 헤더 또는 쿼리로 평문을 함께 보내야 한다.\n\n- 404: 토큰이 존재하지 않음\n- 410: 만료(expires_at) 또는 revoke 됨\n- 401: 비밀번호가 설정된 링크인데 누락/불일치\n- 200: 성공 시 `view_count` 가 1 증가",
        "operationId": "read_shared_document_api_v1_share__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "password",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Password"
            }
          },
          {
            "in": "header",
            "name": "X-Share-Password",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Share-Password"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Shared Document Api V1 Share  Token  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "공개 공유 링크로 문서 조회 (인증 불필요)",
        "tags": [
          "sharing"
        ]
      }
    },
...
  },
...
}